+++ src/modules/auth/decorators/current-user.decorator.edge-cases.spec.ts
@@ -1,0 +1,167 @@
+import { ExecutionContext } from '@nestjs/common';
+import { User } from '../../users/entities/user.entity';
+
+/**
+ * Unit tests for CurrentUser decorator - Edge cases and integration scenarios
+ * Tests handling of null/undefined values and real-world usage scenarios
+ */
+describe('CurrentUser Decorator - Edge Cases & Integration', () => {
+  // Mock user object
+  const mockUser: Partial<User> = {
+    id: '123e4567-e89b-12d3-a456-426614174000',
+    email: 'test@example.com',
+    firstName: 'John',
+    lastName: 'Doe',
+    isActive: true,
+    phoneNumber: '+1234567890',
+    createdAt: new Date('2023-01-01'),
+    updatedAt: new Date('2023-01-01'),
+  };
+
+  // Mock execution context
+  const createMockExecutionContext = (user: Partial<User> | null): ExecutionContext => {
+    const mockRequest = { user };
+    return {
+      switchToHttp: jest.fn().mockReturnValue({
+        getRequest: jest.fn().mockReturnValue(mockRequest),
+      }),
+      getClass: jest.fn(),
+      getHandler: jest.fn(),
+    } as unknown as ExecutionContext;
+  };
+
+  // Simulate the decorator logic for testing
+  const simulateDecoratorLogic = (
+    data: keyof User | undefined,
+    ctx: ExecutionContext,
+  ): User | User[keyof User] => {
+    const request = ctx.switchToHttp().getRequest();
+    const user = request.user;
+    return data ? user?.[data] : user;
+  };
+
+  describe('Edge Cases', () => {
+    it('should return undefined when user is not present in request', () => {
+      // Arrange
+      const context = createMockExecutionContext(null);
+
+      // Act
+      const result = simulateDecoratorLogic(undefined, context);
+
+      // Assert
+      expect(result).toBeNull();
+    });
+
+    it('should return undefined when accessing property of non-existent user', () => {
+      // Arrange
+      const context = createMockExecutionContext(null);
+
+      // Act
+      const result = simulateDecoratorLogic('email', context);
+
+      // Assert
+      expect(result).toBeUndefined();
+    });
+
+    it('should return undefined when accessing non-existent property', () => {
+      // Arrange
+      const context = createMockExecutionContext(mockUser);
+
+      // Act
+      const result = simulateDecoratorLogic('nonExistentProperty' as keyof User, context);
+
+      // Assert
+      expect(result).toBeUndefined();
+    });
+
+    it('should handle user with undefined properties', () => {
+      // Arrange
+      const userWithUndefinedProps: Partial<User> = {
+        id: '123',
+        email: 'test@example.com',
+        firstName: undefined,
+      };
+      const context = createMockExecutionContext(userWithUndefinedProps);
+
+      // Act
+      const result = simulateDecoratorLogic('firstName', context);
+
+      // Assert
+      expect(result).toBeUndefined();
+    });
+
+    it('should handle user with null properties', () => {
+      // Arrange
+      const userWithNullProps: Partial<User> = {
+        id: '123',
+        email: 'test@example.com',
+        phoneNumber: null as unknown as string,
+      };
+      const context = createMockExecutionContext(userWithNullProps);
+
+      // Act
+      const result = simulateDecoratorLogic('phoneNumber', context);
+
+      // Assert
+      expect(result).toBeNull();
+    });
+  });
+
+  describe('Integration Scenarios', () => {
+    it('should work in authenticated request scenario', () => {
+      // Arrange - Simulate JWT guard populating request.user
+      const authenticatedUser: Partial<User> = {
+        id: 'auth-user-id',
+        email: 'authenticated@example.com',
+        firstName: 'Auth',
+        lastName: 'User',
+        isActive: true,
+      };
+      const context = createMockExecutionContext(authenticatedUser);
+
+      // Act
+      const result = simulateDecoratorLogic(undefined, context);
+
+      // Assert
+      expect(result).toEqual(authenticatedUser);
+      expect(result).toHaveProperty('id', 'auth-user-id');
+      expect(result).toHaveProperty('isActive', true);
+    });
+
+    it('should work when extracting only user ID for authorization', () => {
+      // Arrange - Common use case: extract user ID for ownership checks
+      const context = createMockExecutionContext(mockUser);
+
+      // Act
+      const userId = simulateDecoratorLogic('id', context);
+
+      // Assert
+      expect(userId).toBe('123e4567-e89b-12d3-a456-426614174000');
+      expect(typeof userId).toBe('string');
+    });
+
+    it('should handle decorator without parentheses syntax', () => {
+      // Arrange - Testing @CurrentUser() vs @CurrentUser
+      const context = createMockExecutionContext(mockUser);
+
+      // Act
+      const result = simulateDecoratorLogic(undefined, context);
+
+      // Assert
+      expect(result).toBeDefined();
+      expect(result).toEqual(mockUser);
+    });
+
+    it('should handle decorator with property syntax', () => {
+      // Arrange - Testing @CurrentUser('email')
+      const context = createMockExecutionContext(mockUser);
+
+      // Act
+      const result = simulateDecoratorLogic('email', context);
+
+      // Assert
+      expect(result).toBeDefined();
+      expect(result).toBe('test@example.com');
+    });
+  });
+});
