--- src/modules/auth/decorators/current-user.decorator.spec.ts
@@ -1,386 +1,0 @@
-import { ExecutionContext } from '@nestjs/common';
-import { User } from '../../users/entities/user.entity';
-
-/**
- * Unit tests for CurrentUser decorator
- * Tests extraction of user data from request context
- * Coverage target: 95%+
- *
- * Note: Testing param decorators by simulating their logic
- * since NestJS decorators are metadata-based and complex to test directly
- */
-describe('CurrentUser Decorator Logic', () => {
-  // Mock user object
-  const mockUser: Partial<User> = {
-    id: '123e4567-e89b-12d3-a456-426614174000',
-    email: 'test@example.com',
-    firstName: 'John',
-    lastName: 'Doe',
-    isActive: true,
-    phoneNumber: '+1234567890',
-    createdAt: new Date('2023-01-01'),
-    updatedAt: new Date('2023-01-01'),
-  };
-
-  // Mock execution context
-  const createMockExecutionContext = (user: Partial<User> | null): ExecutionContext => {
-    const mockRequest = {
-      user,
-    };
-
-    return {
-      switchToHttp: jest.fn().mockReturnValue({
-        getRequest: jest.fn().mockReturnValue(mockRequest),
-      }),
-      getClass: jest.fn(),
-      getHandler: jest.fn(),
-    } as unknown as ExecutionContext;
-  };
-
-  // Simulate the decorator logic for testing
-  const simulateDecoratorLogic = (
-    data: keyof User | undefined,
-    ctx: ExecutionContext,
-  ): User | User[keyof User] => {
-    const request = ctx.switchToHttp().getRequest();
-    const user = request.user;
-    return data ? user?.[data] : user;
-  };
-
-  describe('Basic Functionality', () => {
-    it('should return entire user object when no data parameter provided', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator(undefined, context);
-
-      // Assert
-      expect(result).toEqual(mockUser);
-      expect(result).toHaveProperty('id');
-      expect(result).toHaveProperty('email');
-      expect(result).toHaveProperty('firstName');
-    });
-
-    it('should extract user from HTTP request context', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator(undefined, context);
-
-      // Assert
-      expect(context.switchToHttp).toHaveBeenCalled();
-      expect(result).toEqual(mockUser);
-    });
-
-    it('should be called with ExecutionContext parameter', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      decorator(undefined, context);
-
-      // Assert
-      expect(context.switchToHttp).toHaveBeenCalledTimes(1);
-    });
-  });
-
-  describe('Property Extraction', () => {
-    it('should return specific user property when data parameter provided', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('email', context);
-
-      // Assert
-      expect(result).toBe('test@example.com');
-      expect(typeof result).toBe('string');
-    });
-
-    it('should extract user id when data is "id"', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('id', context);
-
-      // Assert
-      expect(result).toBe('123e4567-e89b-12d3-a456-426614174000');
-    });
-
-    it('should extract user firstName when data is "firstName"', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('firstName', context);
-
-      // Assert
-      expect(result).toBe('John');
-    });
-
-    it('should extract user lastName when data is "lastName"', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('lastName', context);
-
-      // Assert
-      expect(result).toBe('Doe');
-    });
-
-    it('should extract user isActive status when data is "isActive"', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('isActive', context);
-
-      // Assert
-      expect(result).toBe(true);
-      expect(typeof result).toBe('boolean');
-    });
-
-    it('should extract user createdAt when data is "createdAt"', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('createdAt', context);
-
-      // Assert
-      expect(result).toBeInstanceOf(Date);
-      expect(result).toEqual(new Date('2023-01-01'));
-    });
-
-    it('should extract phoneNumber when data is "phoneNumber"', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('phoneNumber', context);
-
-      // Assert
-      expect(result).toBe('+1234567890');
-    });
-  });
-
-  describe('Edge Cases', () => {
-    it('should return undefined when user is not present in request', () => {
-      // Arrange
-      const context = createMockExecutionContext(null);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator(undefined, context);
-
-      // Assert
-      expect(result).toBeNull();
-    });
-
-    it('should return undefined when accessing property of non-existent user', () => {
-      // Arrange
-      const context = createMockExecutionContext(null);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('email', context);
-
-      // Assert
-      expect(result).toBeUndefined();
-    });
-
-    it('should return undefined when accessing non-existent property', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('nonExistentProperty' as keyof User, context);
-
-      // Assert
-      expect(result).toBeUndefined();
-    });
-
-    it('should handle user with undefined properties', () => {
-      // Arrange
-      const userWithUndefinedProps: Partial<User> = {
-        id: '123',
-        email: 'test@example.com',
-        firstName: undefined,
-      };
-      const context = createMockExecutionContext(userWithUndefinedProps);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('firstName', context);
-
-      // Assert
-      expect(result).toBeUndefined();
-    });
-
-    it('should handle user with null properties', () => {
-      // Arrange
-      const userWithNullProps: Partial<User> = {
-        id: '123',
-        email: 'test@example.com',
-        phoneNumber: null as unknown as string,
-      };
-      const context = createMockExecutionContext(userWithNullProps);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('phoneNumber', context);
-
-      // Assert
-      expect(result).toBeNull();
-    });
-  });
-
-  describe('Type Safety', () => {
-    it('should work with keyof User type parameter', () => {
-      // Arrange
-      const context = createMockExecutionContext(mockUser);
-      const validKey: keyof User = 'email';
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator(validKey, context);
-
-      // Assert
-      expect(result).toBe('test@example.com');
-    });
-
-    it('should handle different user property types correctly', () => {
-      // Arrange
-      const fullUser: Partial<User> = {
-        id: 'uuid-string',
-        email: 'email@test.com',
-        isActive: true,
-        createdAt: new Date(),
-      };
-      const context = createMockExecutionContext(fullUser);
-
-      // Act
-      const stringProp = simulateDecoratorLogic('email', context);
-      const boolProp = simulateDecoratorLogic('isActive', context);
-      const dateProp = simulateDecoratorLogic('createdAt', context);
-
-      // Assert
-      expect(typeof stringProp).toBe('string');
-      expect(typeof boolProp).toBe('boolean');
-      expect(dateProp).toBeInstanceOf(Date);
-    });
-  });
-
-  describe('Integration Scenarios', () => {
-    it('should work in authenticated request scenario', () => {
-      // Arrange - Simulate JWT guard populating request.user
-      const authenticatedUser: Partial<User> = {
-        id: 'auth-user-id',
-        email: 'authenticated@example.com',
-        firstName: 'Auth',
-        lastName: 'User',
-        isActive: true,
-      };
-      const context = createMockExecutionContext(authenticatedUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator(undefined, context);
-
-      // Assert
-      expect(result).toEqual(authenticatedUser);
-      expect(result).toHaveProperty('id', 'auth-user-id');
-      expect(result).toHaveProperty('isActive', true);
-    });
-
-    it('should work when extracting only user ID for authorization', () => {
-      // Arrange - Common use case: extract user ID for ownership checks
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const userId = decorator('id', context);
-
-      // Assert
-      expect(userId).toBe('123e4567-e89b-12d3-a456-426614174000');
-      expect(typeof userId).toBe('string');
-    });
-
-    it('should handle decorator without parentheses syntax', () => {
-      // Arrange - Testing @CurrentUser() vs @CurrentUser
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator(undefined, context);
-
-      // Assert
-      expect(result).toBeDefined();
-      expect(result).toEqual(mockUser);
-    });
-
-    it('should handle decorator with property syntax', () => {
-      // Arrange - Testing @CurrentUser('email')
-      const context = createMockExecutionContext(mockUser);
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      const result = decorator('email', context);
-
-      // Assert
-      expect(result).toBeDefined();
-      expect(result).toBe('test@example.com');
-    });
-  });
-
-  describe('Context Switching', () => {
-    it('should call switchToHttp on execution context', () => {
-      // Arrange
-      const switchToHttpSpy = jest.fn().mockReturnValue({
-        getRequest: jest.fn().mockReturnValue({ user: mockUser }),
-      });
-      const context = {
-        switchToHttp: switchToHttpSpy,
-      } as unknown as ExecutionContext;
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      decorator(undefined, context);
-
-      // Assert
-      expect(switchToHttpSpy).toHaveBeenCalledTimes(1);
-    });
-
-    it('should call getRequest after switching to HTTP context', () => {
-      // Arrange
-      const getRequestSpy = jest.fn().mockReturnValue({ user: mockUser });
-      const context = {
-        switchToHttp: jest.fn().mockReturnValue({
-          getRequest: getRequestSpy,
-        }),
-      } as unknown as ExecutionContext;
-      const decorator = simulateDecoratorLogic;
-
-      // Act
-      decorator(undefined, context);
-
-      // Assert
-      expect(getRequestSpy).toHaveBeenCalledTimes(1);
-    });
-  });
-});
